Simplifies Unspent and moves parsing to CashTokenOutput
